make transforms a folder
